Rename output targets to match component.json.
