+++ lib/conduitjs.js
